+++ lib/src/queue_type.dart
@@ -1,13 +1,13 @@
 part of "alexaquery_types.dart";
 
 class Queue {
-  final InfoText? infoText;
-  final MainArt? mainArt;
-  final MediaReference? mediaReference;
-  final String? playerState;
-  final Progress? progress;
-  final Provider? provider;
-  final String? timestamp;
+  InfoText? infoText;
+  MainArt? mainArt;
+  MediaReference? mediaReference;
+  String? playerState;
+  Progress? progress;
+  Provider? provider;
+  DateTime? timestamp;
 
   Queue({this.infoText, this.mainArt, this.mediaReference, this.playerState, this.progress, this.provider, this.timestamp});
 
@@ -18,22 +18,12 @@ class Queue {
         playerState: json["playerState"],
         progress: json["progress"] == null ? null : Progress.fromJson(json["progress"]),
         provider: json["provider"] == null ? null : Provider.fromJson(json["provider"]),
-        timestamp: timestamp.toIso8601String(),
+        timestamp: timestamp,
       );
-
-  Map<String, dynamic> toJson() => {
-        "infoText": infoText?.toJson(),
-        "mainArt": mainArt?.toJson(),
-        "mediaReference": mediaReference?.toJson(),
-        "playerState": playerState,
-        "progress": progress?.toJson(),
-        "provider": provider?.toJson(),
-        "timestamp": timestamp
-      };
 }
 
 class EndpointList {
-  final Id? id;
+  Id? id;
 
   EndpointList({
     this.id,
@@ -42,16 +32,12 @@ class EndpointList {
   factory EndpointList.fromJson(Map<String, dynamic> json) => EndpointList(
         id: json["id"] == null ? null : Id.fromJson(json["id"]),
       );
-
-  Map<String, dynamic> toJson() => {
-        "id": id?.toJson(),
-      };
 }
 
 class Id {
-  final String? deviceSerialNumber;
-  final String? deviceType;
-  final dynamic softwareVersion;
+  String? deviceSerialNumber;
+  String? deviceType;
+  dynamic softwareVersion;
 
   Id({
     this.deviceSerialNumber,
@@ -64,18 +50,12 @@ class Id {
         deviceType: json["deviceType"],
         softwareVersion: json["softwareVersion"],
       );
-
-  Map<String, dynamic> toJson() => {
-        "deviceSerialNumber": deviceSerialNumber,
-        "deviceType": deviceType,
-        "softwareVersion": softwareVersion,
-      };
 }
 
 class InfoText {
-  final String? subText1;
-  final String? subText2;
-  final String? title;
+  String? subText1;
+  String? subText2;
+  String? title;
 
   InfoText({
     this.subText1,
@@ -88,16 +68,10 @@ class InfoText {
         subText2: json["subText2"],
         title: json["title"],
       );
-
-  Map<String, dynamic> toJson() => {
-        "subText1": subText1,
-        "subText2": subText2,
-        "title": title,
-      };
 }
 
 class MainArt {
-  final String? fullUrl;
+  String? fullUrl;
 
   MainArt({
     this.fullUrl,
@@ -106,14 +80,10 @@ class MainArt {
   factory MainArt.fromJson(Map<String, dynamic> json) => MainArt(
         fullUrl: json["fullUrl"],
       );
-
-  Map<String, dynamic> toJson() => {
-        "fullUrl": fullUrl,
-      };
 }
 
 class MediaReference {
-  final String? value;
+  String? value;
 
   MediaReference({
     this.value,
@@ -122,15 +92,11 @@ class MediaReference {
   factory MediaReference.fromJson(Map<String, dynamic> json) => MediaReference(
         value: json["value"],
       );
-
-  Map<String, dynamic> toJson() => {
-        "value": value,
-      };
 }
 
 class Progress {
-  final int? mediaLength;
-  final int? mediaProgress;
+  int? mediaLength;
+  int? mediaProgress;
 
   Progress({
     this.mediaLength,
@@ -141,16 +107,11 @@ class Progress {
         mediaLength: json["mediaLength"],
         mediaProgress: json["mediaProgress"],
       );
-
-  Map<String, dynamic> toJson() => {
-        "mediaLength": mediaLength,
-        "mediaProgress": mediaProgress,
-      };
 }
 
 class Provider {
-  final ProviderLogo? providerLogo;
-  final String? providerName;
+  ProviderLogo? providerLogo;
+  String? providerName;
 
   Provider({
     this.providerLogo,
@@ -161,16 +122,11 @@ class Provider {
         providerLogo: json["providerLogo"] == null ? null : ProviderLogo.fromJson(json["providerLogo"]),
         providerName: json["providerName"],
       );
-
-  Map<String, dynamic> toJson() => {
-        "providerLogo": providerLogo?.toJson(),
-        "providerName": providerName,
-      };
 }
 
 class ProviderLogo {
-  final String? altText;
-  final String? url;
+  String? altText;
+  String? url;
 
   ProviderLogo({
     this.altText,
@@ -181,9 +137,4 @@ class ProviderLogo {
         altText: json["altText"],
         url: json["url"],
       );
-
-  Map<String, dynamic> toJson() => {
-        "altText": altText,
-        "url": url,
-      };
 }